Fixed scaling logic
