+++ web-api/Cloud3DSTKDeployment.Tests/Cloud3DSTKDeployment.Tests.csproj
@@ -113,7 +113,6 @@
   <ItemGroup>
     <Compile Include="Functional\Cloud3DSTKApiFunctionalTests.cs" />
     <Compile Include="Helpers\ControllerExtensions.cs" />
-    <Compile Include="Models\CreateApiJsonBody.cs" />
     <Compile Include="Models\DeletePoolApiJsonBody.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Unit\Cloud3DSTKApiUnitTests.cs" />
