+++ web-api/Cloud3DSTKDeployment/Cloud3DSTKDeployment.csproj
@@ -147,6 +147,7 @@
     </Compile>
     <Compile Include="Helpers\ConfigurationHelper.cs" />
     <Compile Include="Models\ApiResultMessages.cs" />
+    <Compile Include="Models\CreateApiJsonBody.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\BatchService.cs" />
     <Compile Include="Services\IBatchService.cs" />
